Cleanup: Use more intermediate variables
It's in general a very bad idea to overwrite a local variable with new
content. I promise. Try avoiding it, you'll like it. It's much easier to
debug code written that way, since it means the result of previous
calculations/before calculations are available, which can be incredibly
helpful in trying to understand _why_ strange things are happening.